requirements.txt
chromadb
sentence-transformers
torch
Pillow
pypdf
scikit-learn
